[platform-api] overriding default weak otPlatTimeGet (#12)
Later OpenThread added a default weak implementation of otPlatTimeGet
API in radio_platform.cpp. But the strong implementation in
ot-esp32/src/time.c failed to override the default one. This is
because there is no other unresolved symbols in src/time.c and the
linker will discard the entire .o file.

This commit moves otPlatTimeGet to a file includes other unresolved
symbols, and src/alarm.c seems to be a good candidate (since they are
both about time).